+++ interpreterPatternPresentation/interpreterdemo/src/main/java/edu/umich/bolle/Expression.java
@@ -1,5 +1,6 @@
 package edu.umich.bolle;
 
 public interface Expression {
-    int interpret(Context context);
+    // root abstraction for interpreter pattern
+    double interpret(Context context);
 }
